Updated Memory Routes

+++ server/src/controllers/memoryController.ts
@@ -8,23 +8,20 @@ import CustomRequest from "../types/CustomRequest";
 const REQ_URL = "https://api.mem0.ai/v1/memories/";
 const TOKEN = process.env.MEM0_TOKEN as string;
 
-const getMemory = async (req: Request, res: Response) => {
+const getMemory = async (req: CustomRequest, res: Response) => {
   let success = false;
-
-  // Saving req data into a variable
-  let { user_id } = req.query;
 
   try {
     // check if user exists
-      // let user = await User.findById(req.user.id);
-      // if (!user) {
-      //   return res.status(400).json({ success, error: "User not found!" });
-      // }
+    let user = await User.findById(req.user.id);
+    if (!user) {
+      return res.status(400).json({ success, error: "User not found!" });
+    }
     const options = {
       method: "GET",
       headers: { Authorization: TOKEN },
     };
-    const response = await fetch(`${REQ_URL}?user_id=${user_id}`, options);
+    const response = await fetch(`${REQ_URL}?user_id=${user.userId}`, options);
     const data = await response.json();
 
     success = true;
@@ -35,15 +32,21 @@ const getMemory = async (req: Request, res: Response) => {
   }
 };
 
-const addMemory = async (req: Request, res: Response) => {
+const addMemory = async (req: CustomRequest, res: Response) => {
   let success = false;
-  let { user_id, messages } = req.body;
+  let { messages } = req.body;
   try {
+    // check if user exists
+    let user = await User.findById(req.user.id);
+    if (!user) {
+      return res.status(400).json({ success, error: "User not found!" });
+    }
+
     const options = {
       method: "POST",
       headers: { Authorization: TOKEN, "Content-Type": "application/json" },
       body: JSON.stringify({
-        user_id: user_id,
+        user_id: user.userId,
         messages: messages,
       }),
     };
@@ -60,10 +63,16 @@ const addMemory = async (req: Request, res: Response) => {
   } catch (error) {}
 };
 
-const deleteMemory = async (req: Request, res: Response) => {
+const deleteMemory = async (req: CustomRequest, res: Response) => {
   let success = false;
   let { memoryid } = req.query;
   try {
+    // check if user exists
+    let user = await User.findById(req.user.id);
+    if (!user) {
+      return res.status(400).json({ success, error: "User not found!" });
+    }
+
     const options = {
       method: "DELETE",
       headers: { Authorization: TOKEN, "Content-Type": "application/json" },
@@ -87,61 +96,70 @@ const deleteMemory = async (req: Request, res: Response) => {
   }
 };
 
-const updateMemory = async (req: Request, res: Response) => {
-    let success = false;
-    let { memoryid, text } = req.body;
+const updateMemory = async (req: CustomRequest, res: Response) => {
+  let success = false;
+  let { memoryid, text } = req.body;
 
-    try {
-      const options = {
-        method: "PUT",
-        headers: { Authorization: TOKEN, "Content-Type": "application/json" },
-        body: JSON.stringify({
-            text: text
-        })
-      };
-      const response = await fetch(
-        `https://api.mem0.ai/v1/memories/${memoryid}/`,
-        options
-      );
-      const data = await response.json();
-  
-      if (data) {
-        success = true;
-        return res.json({ success, memories: data });
-      } else {
-        success = false;
-        return res.json({ success });
-      }
-    } catch (error) {
-      console.log(error);
-      return res.status(500).json({ success, error: "Internal Server Error!" });
+  try {
+    // check if user exists
+    let user = await User.findById(req.user.id);
+    if (!user) {
+      return res.status(400).json({ success, error: "User not found!" });
     }
-  };
 
-  const searchMemory = async (req: Request, res: Response) => {
-    let success = false;
-    let { user_id, query } = req.body;
-    try {
-      const options = {
-        method: "POST",
-        headers: { Authorization: TOKEN, "Content-Type": "application/json" },
-        body: JSON.stringify({
-          user_id: user_id,
-          query: query,
-          top_k: 5
-        }),
-      };
-      const response = await fetch(`https://api.mem0.ai/v1/memories/search/`, options);
-      const data = await response.json();
+    const options = {
+      method: "PUT",
+      headers: { Authorization: TOKEN, "Content-Type": "application/json" },
+      body: JSON.stringify({
+        text: text,
+      }),
+    };
+    const response = await fetch(
+      `https://api.mem0.ai/v1/memories/${memoryid}/`,
+      options
+    );
+    const data = await response.json();
 
-      if (data) {
-        success = true;
-        return res.json({ success, memories: data });
-      } else {
-        success = false;
-        return res.json({ success });
-      }
-    } catch (error) {}
-  };
+    if (data) {
+      success = true;
+      return res.json({ success, memories: data });
+    } else {
+      success = false;
+      return res.json({ success });
+    }
+  } catch (error) {
+    console.log(error);
+    return res.status(500).json({ success, error: "Internal Server Error!" });
+  }
+};
+
+const searchMemory = async (req: Request, res: Response) => {
+  let success = false;
+  let { user_id, query } = req.body;
+  try {
+    const options = {
+      method: "POST",
+      headers: { Authorization: TOKEN, "Content-Type": "application/json" },
+      body: JSON.stringify({
+        user_id: user_id,
+        query: query,
+        top_k: 5,
+      }),
+    };
+    const response = await fetch(
+      `https://api.mem0.ai/v1/memories/search/`,
+      options
+    );
+    const data = await response.json();
+
+    if (data) {
+      success = true;
+      return res.json({ success, memories: data });
+    } else {
+      success = false;
+      return res.json({ success });
+    }
+  } catch (error) {}
+};
 
 export { getMemory, addMemory, deleteMemory, updateMemory, searchMemory };
